@@ -1,30 +1,31 @@
 {
-  "name": "pbotool-ce-sample",
+  "name": "fsm--msf--updated",
   "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite --host --port 8080",
     "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@heroicons/vue": "^2.1.4",
-    "chart.js": "^4.4.3",
+    "@heroicons/vue": "^2.1.5",
+    "chart.js": "^4.4.5",
     "js-yaml": "^4.1.0",
-    "marked": "^13.0.1",
-    "pinia": "^2.1.7",
-    "read-excel-file": "^5.8.3",
+    "marked": "^14.1.3",
+    "pinia": "^2.2.4",
+    "read-excel-file": "^5.8.6",
     "tippy.js": "^6.3.7",
     "uuid": "^10.0.0",
-    "vue": "^3.4.30",
+    "vue": "^3.5.10",
     "vue-chartjs": "^5.3.1"
   },
   "devDependencies": {
-    "@tailwindcss/typography": "^0.5.13",
-    "@vitejs/plugin-vue": "^5.0.5",
-    "autoprefixer": "^10.4.19",
-    "postcss": "^8.4.38",
-    "tailwindcss": "^3.4.4",
-    "vite": "^5.3.1"
+    "@tailwindcss/typography": "^0.5.15",
+    "@vitejs/plugin-vue": "^5.1.4",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.14",
+    "vite": "^5.4.8"
   }
 }
